requirements.txt
# FastAPI Backend
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Streamlit Frontend
streamlit==1.28.1

# Environment
python-dotenv==1.0.0

# ML & Embeddings
sentence-transformers==2.2.2
torch==2.1.0
transformers==4.35.0

# PDF Processing
pypdf==3.17.0
pdfplumber==0.10.3

# Vector Databases
chromadb==0.4.18
pinecone-client==2.2.4

# HTTP Requests
requests==2.31.0
httpx==0.25.1

# Utilities
numpy==1.24.3
pandas==2.1.3

# Logging
colorlog==6.8.0